// features/dashboard/types/admin-meta.ts
export type FieldMeta = {
  path: string
  label: string
  description: string | null
  fieldMeta: JSONValue | null

  viewsIndex: number
  customViewsIndex: number | null
  views: FieldViews[number]
  controller: FieldController<unknown, JSONValue>

  search: 'default' | 'insensitive' | null
  graphql: {
    isNonNull: ('read' | 'create' | 'update')[]
  }
  createView: {
    fieldMode: 'edit' | 'hidden'
  }
  itemView: {
    fieldMode: 'edit' | 'read' | 'hidden'
    fieldPosition: 'form' | 'sidebar'
  }
  listView: {
    fieldMode: 'read' | 'hidden'
  }

  isFilterable: boolean
  isOrderable: boolean
}

export type FieldGroupMeta = {
  label: string
  description: string | null
  fields: FieldMeta[]
}

export type ListMeta = {
  key: string
  path: string
  description: string | null

  // List naming
  label: string
  labelField: string
  singular: string
  plural: string
  itemQueryName: string
  listQueryName: string

  // Fields and groups
  fields: { [path: string]: FieldMeta }
  groups: FieldGroupMeta[]
  
  // GraphQL names
  graphql: {
    names: GraphQLNames
  }
  gqlNames?: GraphQLNames // For backward compatibility

  // List configuration
  pageSize: number
  initialColumns: string[]
  initialSearchFields: string[]
  initialSort: null | { direction: 'ASC' | 'DESC'; field: string }
  isSingleton: boolean
  isHidden?: boolean

  // UI configuration
  hideNavigation: boolean
  hideCreate: boolean
  hideDelete: boolean
}

export type AdminMeta = {
  lists: { [list: string]: ListMeta }
  listsByPath: { [path: string]: ListMeta }
}